updated home screen so that if user doesn't have saved recipes/inventory, it doesn't display and changed screen names to include 'screen'
